remove workarounds
